move code-generation to own pre-compiled script plugin
