+++ gradle/code-generation/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+	id("build-logic.kotlin-dsl-gradle-plugin")
+}
+
+allprojects {
+	group = "com.tegonal.minimalist.code-generation"
+}
+
+dependencies {
+	api(buildLibs.kbox)
+}
